add LGPL Shape Healing code from SALOME GEOM compiles, but not yet used
git-svn-id: https://cam-occ.googlecode.com/svn@120 2359b6c2-501e-0410-8cfe-a57835d51191

+++ trunk/src/uio/nanotimer.cc
@@ -58,7 +58,7 @@ std::string nanotimer::humanreadable(double s) {
   int m=0;
   if (s > 60) {
     m = s/60;
-    s = s-(m*60);
+    s = s-(double)(m*60);
   }
   std::string out = uio::toString(m) + "m, " + uio::toString(s) + "s";
   return out;
